Make streaming_source_manager impl header-only
